% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect.R
\name{classify_sheet}
\alias{classify_sheet}
\alias{split_sheet}
\title{Classify and split sheet}
\usage{
classify_sheet(sheet, as = "limits")

split_sheet(sheet)
}
\arguments{
\item{sheet}{A linen \code{worksheet} object, possibly from an
Excel or googlesheets spreadsheet.}

\item{as}{Character, indicating what to return - \code{"limits"}:
a list of limits (the default), \code{"groups"} a matrix of the
same dimensions as the worksheet indicating what group each cell
is in, or \code{"both"}: a list with elements \code{"limits"}
and \code{"groups"}.}
}
\description{
Classify a table into sub-regions.  We're looking for a (possibly
ragged) block of cells surrounded by a set of blank cells or the
edge of the sheet.  This does not detect regions that are multiple
square regions offset from each other, but that could possibly be
done afterwards.
}
\details{
This function works by applying the "flood fill" algorithm to
non-blank cells in the worksheet and then squaring off the result.

The \code{classify_sheet} function does the actual classification,
and \code{split_sheet} applies \code{worksheet_view} to these to
produce something that can be used (approximately) as if it was a
separate sheet.
}

