% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{jailbreak_read}
\alias{jailbreak_read}
\title{Read an xlsx file that probably contains nontabular data}
\usage{
jailbreak_read(path, sheet = 1L)
}
\arguments{
\item{path}{Path to the xlsx file to load.  xls files are not supported.}

\item{sheet}{Sheet number (not name at this point).  Googlesheets
exported sheets are likely not to do the right thing.}
}
\value{
An \code{xlsx} object that can be printed.  Future methods
  might do something sensible.  The structure is subject to
  complete change so is not documented here.
}
\description{
This function does not get the data into a usable form but at
least loads it up into R so we can poke about with it.  Dates are
not yet handled because they're complicated.  The resulting loaded
data can distinguish between formulae and data, numbers and text.
Merged cells are detected.  A summary of the data is printed if
you print the resulting object.
}

