% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readxl.R
\name{jailbreak_readxl}
\alias{jailbreak_readxl}
\title{Read an Excel spreadsheet like readxl}
\usage{
jailbreak_readxl(path, sheet = 1L, col_names = TRUE, col_types = NULL,
  na = "", skip = 0)
}
\arguments{
\item{path}{Path to the xlsx file}

\item{sheet}{Sheet name or an integer}

\item{col_names}{TRUE (the default) indicating we should use the
first row as column names, FALSE, indicating we should generate
names (X1, X2, ..., Xn) or a character vector of names to apply.}

\item{col_types}{Either NULL (the default) indicating we should
guess the column types or a vector of column types (values must
be "blank", "numeric", "date" or "text").}

\item{na}{Values indicating missing values (if different from
blank).  Not yet used.}

\item{skip}{Number of rows to skip.}
}
\description{
Read an Excel spreadsheet the same way as readxl, but slower.
Assumes a well behaved table of data.
}

