% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{jailbreak_subset}
\alias{jailbreak_subset}
\title{Subset a jailbreakr sheet}
\usage{
jailbreak_subset(dat, xr)
}
\arguments{
\item{dat}{A jailbreakr sheet}

\item{xr}{A \code{cell_limits} object indicating the extent to
extract.}
}
\description{
Subset a jailbreakr sheet based on a rectangular selection based
on a cellranger \code{cell_limits} object.
}

