% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unmerge_headers.R
\name{unmerge_header}
\alias{unmerge_header}
\title{Unmerge headers and row labels}
\usage{
unmerge_header(sheet, xr, horizontal = TRUE, sep = ":")
}
\arguments{
\item{sheet}{A \code{\link{worksheet}} object}

\item{xr}{A cell range (a \code{cellranger::cell_limits} object)
indicating the region to collapse.}

\item{horizontal}{Flag indicating if this a horizontal region
representing headers (\code{TRUE}, the default) or a vertical
region.  If horizontal we collapse away vertical spaces to
create a single header row.  If vertical, we collapse away
horizontal space to create a single row names column.}

\item{sep}{The separator to use between collapsed elements.  The
default is a colon (\code{:}).  Because of the sheets this is
designed to work with, you're never going to get syntactically
valid names here, so feel free to use anything.  This can be
multiple characters, a newline, whatever.  Likely you're going
to have to process these names a bit later.}
}
\value{
A character vector
}
\description{
Unmerge cells that represent hierarchical headers and row labels.
}
\details{
There is a pattern in headers where we have some number of levels
of merge, most commonly:

\preformatted{
  | -------X------- |
  |  a  |  b  |  c  |
}

which typically is intended to be interpreted as:

\preformatted{
  | X:a | X:b | X:c |
}

The same thing happen on rows, too.  I suspect that this is
generalisable to more than 2 columns.

Within such header rows are either: vertically merged:

\preformatted{
  | X | ===> | X |
  | | |
}

or stacked

\preformatted{
  | X | ===> | X:a |
  | a |
}

or stacked and blank

\preformatted{
  | X | or |   |  ===> | X |
  |   |    | X |
}

I want to write something that handles this, that can be hooked up
to act as the "header" section on a view.  For now, let's apply it
to all cells within a range.
}

