% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_metadata.R
\name{split_metadata}
\alias{split_metadata}
\title{Split metadata from a worksheet}
\usage{
split_metadata(x, include_merged = TRUE, min_jump = 2, min_data_block = 5)
}
\arguments{
\item{sheet}{A worksheet or a worksheet view}
}
\value{
A worksheet view
}
\description{
Split metadata from a worksheet or region of a worksheet (a
worksheet view).
}
\details{
There is a common pattern in spreadsheets where data is stored as:

\preformatted{
mmm
mmm
HHHHHHHH
dddddddd
dddddddd
}

where \code{m} is some metadata (perhaps indicating table name,
creator, dates, etc), \code{H} is the header and \code{d} is the
actual data.  This function will split the metadata (\code{m})
part off, leaving a table that is more suitable for further
processing.  In many ways this is like the \code{skip} argument to
\code{readxl::read_excel} and \code{read.csv}, but we will retain
the metadata (somewhere!).

The idea here is that the metadata block starts when we get a
shift in the number of non-blank cells.  There needs to be some
heuristics here to help, and things will need to be tuneable: do
merged cells count as non-empty (\code{include_merged}; default is
to include them), how big a jump we look for (\code{min_jump};
default is 2 columns), how many rows of the same size do we look
for in the data block (\code{min_data_block}; default is 5 rows).

Other things that might be useful, but which aren't supported yet,
include looking for different colours and fonts in the metadata
and the main block; when we switch from one to the other we're
likely to see things like a change here.
}
\author{
Rich FitzJohn
}

